refactor: move subcomponent props interface to new file
